+++ src/elements.rs
@@ -20,6 +20,7 @@
 //!   - [`PaddedElement`][]: adds a padding to the wrapped element
 //!   - [`StyledElement`][]: sets a default style for the wrapped element and its children
 //! - Other:
+//!   - [`Image`][]: an image (requires the `images` feature)
 //!   - [`Break`][]: adds forced line breaks as a spacer
 //!   - [`PageBreak`][]: adds a forced page break
 //!
@@ -31,12 +32,16 @@
 //! [`OrderedList`]: struct.OrderedList.html
 //! [`UnorderedList`]: struct.UnorderedList.html
 //! [`Text`]: struct.Text.html
+//! [`Image`]: struct.Image.html
 //! [`Break`]: struct.Break.html
 //! [`PageBreak`]: struct.PageBreak.html
 //! [`Paragraph`]: struct.Paragraph.html
 //! [`FramedElement`]: struct.FramedElement.html
 //! [`PaddedElement`]: struct.PaddedElement.html
 //! [`StyledElement`]: struct.StyledElement.html
+
+#[cfg(feature = "images")]
+mod images;
 
 use std::collections;
 use std::iter;
@@ -46,7 +51,10 @@ use crate::error::{Error, ErrorKind};
 use crate::render;
 use crate::style::{Style, StyledString};
 use crate::wrap;
-use crate::{Context, Element, Margins, Mm, Position, RenderResult, Size};
+use crate::{Alignment, Context, Element, Margins, Mm, Position, RenderResult, Size};
+
+#[cfg(feature = "images")]
+pub use images::Image;
 
 /// Arranges a list of elements sequentially.
 ///
@@ -179,27 +187,6 @@ impl Element for Text {
     }
 }
 
-/// The alignment of a [`Paragraph`][].
-///
-/// The default alignment is left-flushed.
-///
-/// [`Paragraph`]: struct.Paragraph.html
-#[derive(Clone, Copy, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
-pub enum Alignment {
-    /// Left-flushed.
-    Left,
-    /// Right-flushed.
-    Right,
-    /// Centered.
-    Center,
-}
-
-impl Default for Alignment {
-    fn default() -> Alignment {
-        Alignment::Left
-    }
-}
-
 /// A multi-line wrapped paragraph of formatted text.
 ///
 /// If the text of this paragraph is longer than the page width, the paragraph is wrapped at word
@@ -223,7 +210,7 @@ impl Default for Alignment {
 /// p.push("This is an ");
 /// p.push_styled("important", style::Color::Rgb(255, 0, 0));
 /// p.push(" message!");
-/// p.set_alignment(elements::Alignment::Center);
+/// p.set_alignment(genpdf::Alignment::Center);
 /// ```
 ///
 /// Chained:
@@ -233,11 +220,11 @@ impl Default for Alignment {
 ///     .string("This is an ")
 ///     .styled_string("important", style::Color::Rgb(255, 0, 0))
 ///     .string(" message!")
-///     .aligned(elements::Alignment::Center);
+///     .aligned(genpdf::Alignment::Center);
 /// ```
 ///
 /// [`Style`]: ../style/struct.Style.html
-/// [`Alignment`]: enum.Alignment.html
+/// [`Alignment`]: ../enum.Alignment.html
 /// [`push`]: #method.push
 /// [`push_styled`]: #method.push_styled
 /// [`string`]: #method.string